requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
blessed=1.20.0=pypi_0
bzip2=1.0.8=h5eee18b_5
ca-certificates=2024.3.11=h06a4308_0
certifi=2024.2.2=pypi_0
charset-normalizer=3.3.2=pypi_0
cmake=3.29.2=pypi_0
contourpy=1.2.1=pypi_0
cudatoolkit=11.8.0=h6a678d5_0
cycler=0.12.1=pypi_0
expat=2.6.2=h6a678d5_0
filelock=3.14.0=pypi_0
fonttools=4.51.0=pypi_0
fsspec=2024.3.1=pypi_0
gpustat=1.1.1=pypi_0
grpcio=1.63.0=pypi_0
h5py=3.10.0=pypi_0
huggingface-hub=0.23.0=pypi_0
idna=3.7=pypi_0
jinja2=3.1.3=pypi_0
joblib=1.4.2=pypi_0
kiwisolver=1.4.5=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
lit=18.1.4=pypi_0
markdown=3.6=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.7.1=pypi_0
mpmath=1.3.0=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.3=pypi_0
numpy=1.24.2=pypi_0
nvidia-cublas-cu11=11.10.3.66=pypi_0
nvidia-cuda-cupti-cu11=11.7.101=pypi_0
nvidia-cuda-nvrtc-cu11=11.7.99=pypi_0
nvidia-cuda-runtime-cu11=11.7.99=pypi_0
nvidia-cudnn-cu11=8.5.0.96=pypi_0
nvidia-cufft-cu11=10.9.0.58=pypi_0
nvidia-curand-cu11=10.2.10.91=pypi_0
nvidia-cusolver-cu11=11.4.0.1=pypi_0
nvidia-cusparse-cu11=11.7.4.91=pypi_0
nvidia-ml-py=12.550.52=pypi_0
nvidia-nccl-cu11=2.14.3=pypi_0
nvidia-nvtx-cu11=11.7.91=pypi_0
opencv-python=4.8.0.74=pypi_0
openssl=3.0.13=h7f8727e_1
packaging=24.0=pypi_0
pandas=2.0.3=pypi_0
pillow=10.2.0=pypi_0
pip=23.3.1=py310h06a4308_0
protobuf=5.26.1=pypi_0
psutil=5.9.8=pypi_0
pyparsing=3.1.2=pypi_0
python=3.10.14=h955ad1f_0
python-dateutil=2.9.0.post0=pypi_0
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=h5eee18b_0
regex=2024.4.28=pypi_0
requests=2.31.0=pypi_0
safetensors=0.4.3=pypi_0
scikit-learn=1.3.0=pypi_0
scipy=1.13.0=pypi_0
setuptools=68.2.2=py310h06a4308_0
six=1.16.0=pypi_0
sqlite=3.41.2=h5eee18b_0
sympy=1.12=pypi_0
tensorboard=2.16.2=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorboardx=2.6.2.2=pypi_0
threadpoolctl=3.5.0=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.15.2=pypi_0
torch=2.0.0=pypi_0
torchvision=0.15.1=pypi_0
tqdm=4.61.2=pypi_0
transformers=4.38.2=pypi_0
triton=2.0.0=pypi_0
typing-extensions=4.11.0=pypi_0
tzdata=2024.1=pypi_0
urllib3=2.2.1=pypi_0
wcwidth=0.2.13=pypi_0
werkzeug=3.0.3=pypi_0
wheel=0.41.2=py310h06a4308_0
xz=5.4.6=h5eee18b_0
zlib=1.2.13=h5eee18b_0
